# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Folder Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /folders/{id}/sets:
    get:
      operationId: listQuestionSets
      tags: [questionSets]
      summary: List question sets in a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of question sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionSet"
  /sets/{id}:
    get:
      operationId: getQuestionSet
      tags: [questionSets]
      summary: Get a question set with its questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question set with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionSetDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuestionSet
      tags: [questionSets]
      summary: Delete a question set and its questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sets/{setId}/link-questions:
    post:
      operationId: linkQuestions
      tags: [questionSets]
      summary: Link existing questions into another set (zero-copy)
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkQuestionsInput"
      responses:
        "200":
          description: Links created
          content:
            application/json:
              schema:
                type: object
                properties:
                  linked:
                    type: integer
                required:
                  - linked
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /links/{id}:
    delete:
      operationId: deleteQuestionLink
      tags: [questionSets]
      summary: Remove a question link (unlink from set, original question untouched)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuestionLink
      tags: [questionSets]
      summary: Update hidden parts for a linked CQ question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdate"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionLink"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    QuestionSet:
      type: object
      required:
        - id
        - folderId
        - name
        - totalQuestions
        - createdAt
      properties:
        id:
          type: integer
        folderId:
          type: integer
        name:
          type: string
        examType:
          type: ["string", "null"]
        totalQuestions:
          type: integer
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    QuestionOption:
      type: object
      required:
        - letter
        - text
      properties:
        letter:
          type: string
        text:
          type: string
    QuestionPart:
      type: object
      required:
        - key
        - label
        - text
      properties:
        key:
          type: string
        label:
          type: string
        text:
          type: string
        solution:
          type: ["string", "null"]
        aiSolution:
          type: ["string", "null"]
    Question:
      type: object
      required:
        - id
        - setId
        - chorchaId
        - questionIndex
        - type
        - questionText
        - options
        - parts
        - stemImages
        - hidden
        - createdAt
      properties:
        id:
          type: integer
        setId:
          type: integer
        chorchaId:
          type: string
        questionIndex:
          type: integer
        type:
          type: string
        questionText:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        parts:
          type: array
          items:
            $ref: "#/components/schemas/QuestionPart"
        answer:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        stemImages:
          type: array
          items:
            type: string
        aiExplanation:
          type: ["string", "null"]
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        linkId:
          type: ["integer", "null"]
          description: Set if this question is linked (not owned) by this set
        hiddenParts:
          type: array
          items:
            type: string
          description: CQ part keys hidden in this set (only for linked questions)
    QuestionLink:
      type: object
      required:
        - id
        - questionId
        - setId
        - questionIndex
        - hiddenParts
        - createdAt
      properties:
        id:
          type: integer
        questionId:
          type: integer
        setId:
          type: integer
        questionIndex:
          type: integer
        hiddenParts:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    LinkQuestionsInput:
      type: object
      required:
        - questionIds
      properties:
        questionIds:
          type: array
          items:
            type: integer
    LinkUpdate:
      type: object
      properties:
        hiddenParts:
          type: array
          items:
            type: string
    QuestionSetDetail:
      type: object
      required:
        - set
        - questions
      properties:
        set:
          $ref: "#/components/schemas/QuestionSet"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string